requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
Deprecated==1.2.13
distro==1.7.0
dj-database-url==0.5.0
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
kombu==5.2.4
packaging==21.3
paramiko==2.11.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
redis==4.3.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
wrapt==1.14.1
